Refactored non-database models to validation file.
